Add rating updates

+++ schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE users (
     id SERIAL PRIMARY KEY,
     username TEXT UNIQUE,
     password TEXT,
-    rating INTEGER
+    rating REAL
 );
 
 CREATE TABLE games (
